+++ .dangerfile/policy_tests.rb
@@ -151,6 +151,7 @@ def policy_bad_indentation?(file)
   fail_message = ""
 
   indent_level = 0
+  code_block = false
   eos_block = false
   define_block = false
 
@@ -161,10 +162,11 @@ def policy_bad_indentation?(file)
     indentation = line.match(/\A\s*/).to_s.length
 
     # Skip blocks of EOS/EOF text and define blocks, since these contain arbitrarily spaced code/text
+    code_block = false if code_block && line.strip == "EOS" || line.strip == "EOF"
     eos_block = false if eos_block && line.strip == "EOS" || line.strip == "EOF"
     define_block = false if define_block && line.start_with?("end")
 
-    if !eos_block && !define_block
+    if !code_block && !define_block && !eos_block
       indent_level -= 2 if line.strip == "end" || line.strip == ")"
 
       if indentation != indent_level && !line.strip.empty? && line.strip != "EOS" && line.strip != "EOF"
@@ -173,11 +175,13 @@ def policy_bad_indentation?(file)
 
       indent_level += 2 if line.strip.end_with?(" do") || line.start_with?("info(")
 
-      eos_block = true if line.include?("<<-")
+      # We only check EOS blocks if they are for the code field
+      code_block = true if line.include?("<<-") && line.include?("code")
+      eos_block = true if line.include?("<<-") && !line.include?("code")
       define_block = true if line.start_with?("define ") && line.strip.end_with?(" do")
 
     # If we're within one of these blocks, at least make sure we're 2 spaces indented
-    elsif (eos_block || define_block) && indentation < 2 && !line.strip.empty?
+    elsif (code_block || define_block) && indentation < 2 && !line.strip.empty?
       fail_message += "Line #{line_number.to_s}: Expected indentation of at least two spaces within code/text block.\n"
     end
   end