+++ tools/meta_parent_policy_compiler/meta_parent_policy_compiler.rb
@@ -6,6 +6,7 @@ require "json"
 # More info at https://github.com/flexera-public/policy_templates/blob/master/README_META_POLICIES.md
 default_child_policy_template_files = [
   # AWS Policy Templates
+  "../../automation/aws/aws_missing_regions/aws_missing_regions.pt",
   "../../compliance/aws/disallowed_regions/aws_disallowed_regions.pt",
   "../../compliance/aws/instances_without_fnm_agent/aws_instances_not_running_flexnet_inventory_agent.pt",
   "../../compliance/aws/long_stopped_instances/aws_long_stopped_instances.pt",