+++ compliance/flexera/fnms/ignored_recent_inventory_dates/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.4
 
 - Fixed spelling errors in short description and `Days since last inventory` parameter.
+- Updated policy metadata to make it more clear what Flexera service the policy is for
 
 ## v2.3